--- api/pkg/aws/configuration.go
@@ -1,5 +1,0 @@
-package aws
-
-type Configuration struct {
-	Region string `long:"region" description:"AWS region to use" default:"eu-north-1"`
-}
