+++ api/pkg/aws/enterprise/cloud/session.go
@@ -1,9 +1,13 @@
-package aws
+package cloud
 
 import (
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/session"
 )
+
+type Configuration struct {
+	Region string `long:"region" description:"AWS region to use" default:"eu-north-1"`
+}
 
 func New(cfg *Configuration) (*session.Session, error) {
 	return session.NewSession(&aws.Config{
