+++ api/pkg/aws/module/cloud.go
@@ -1,0 +1,13 @@
+//go:build cloud
+// +build cloud
+
+package module
+
+import (
+	"getsturdy.com/api/pkg/aws/enterprise/cloud"
+	"getsturdy.com/api/pkg/di"
+)
+
+func Module(c *di.Container) {
+	c.Import(cloud.Module)
+}
