+++ api/pkg/aws/module/not_cloud.go
@@ -1,0 +1,8 @@
+//go:build !cloud
+// +build !cloud
+
+package module
+
+import "getsturdy.com/api/pkg/di"
+
+func Module(c *di.Container) {}
