+++ api/pkg/configuration/enterprise/cloud/configuration.go
@@ -2,9 +2,11 @@ package cloud
 
 import (
 	"getsturdy.com/api/pkg/analytics/enterprise/cloud/posthog"
+	aws "getsturdy.com/api/pkg/aws/enterprise/cloud"
 	"getsturdy.com/api/pkg/configuration"
-	"getsturdy.com/api/pkg/emails/enterprise/cloud"
+	emails "getsturdy.com/api/pkg/emails/enterprise/cloud"
 	"getsturdy.com/api/pkg/github/enterprise/config"
+	queue "getsturdy.com/api/pkg/queue/enterprise/cloud"
 
 	"github.com/jessevdk/go-flags"
 )
@@ -12,9 +14,11 @@ import (
 type Configuration struct {
 	configuration.Base
 
+	AWS       *aws.Configuration      `flags-group:"aws" namespace:"aws"`
 	GitHub    *config.GitHubAppConfig `flags-group:"github-app" namespace:"github-app"`
 	Analytics *posthog.Configuration  `flags-group:"analytics" namespace:"analytics"`
-	Emails    *cloud.Configuration    `flags-group:"emails" namespace:"emails"`
+	Emails    *emails.Configuration   `flags-group:"emails" namespace:"emails"`
+	Queue     *queue.Configuration    `flags-group:"queue" namespace:"queue"`
 }
 
 func New() (Configuration, error) {
