+++ api/pkg/queue/enterprise/cloud/module.go
@@ -1,4 +1,4 @@
-package queue
+package cloud
 
 import "getsturdy.com/api/pkg/di"
