+++ api/pkg/queue/module/cloud.go
@@ -1,0 +1,13 @@
+//go:build cloud
+// +build cloud
+
+package queue
+
+import (
+	"getsturdy.com/api/pkg/di"
+	"getsturdy.com/api/pkg/queue/enterprise/cloud"
+)
+
+func Module(c *di.Container) {
+	c.Import(cloud.Module)
+}
