+++ api/pkg/queue/module/not_cloud.go
@@ -1,0 +1,13 @@
+//go:build !cloud
+// +build !cloud
+
+package queue
+
+import (
+	"getsturdy.com/api/pkg/di"
+	"getsturdy.com/api/pkg/queue"
+)
+
+func Module(c *di.Container) {
+	c.Register(queue.NewInMemory, new(queue.Queue))
+}
